requirements.txt
pandas~=1.5.3
streamlit~=1.21.0
plotly~=5.14.1
binance-connector~=2.0.0